Ajout liens connexion et inscription sur les pages correspondantes
Trie des ressources selon l'ordre alphabétique (ou non) du libelé ou du type de la ressource

Ajout de la localisation de la ressource
Suppression du paramètre du type de moteur de session

Enregistrement des classes dans l'administration

+++ rsc_mgmt/views.py
@@ -1,4 +1,4 @@
-from django.http import HttpResponse, JsonResponse
+from django.http import JsonResponse
 from django.shortcuts import render, redirect
 
 from .forms import *
@@ -8,12 +8,33 @@ def is_login(request):
     return request.session.__contains__('user')
 
 
+# Pourrais être amélioré en utilisant les actions (ex: 'can_create_resource', 'can_view', ...)
+def is_admin(user):
+    return user['is_admin']
+
+
 # Create your views here.
 def index(request):
     if not is_login(request):
         return redirect('login')
 
-    return render(request, 'rsc_mgmt/index.html', {'res': Resource.objects.all(), 'type': ResourceType.objects.all()})
+    order = request.GET.get('order', default='libele')
+    res = None
+
+    """
+    Si on a un signe négatif dans le paramètre de trie, on retire ce signe pour inverser le filtre
+    """
+    if "-" in order:  # Si il y a un paramètre
+        res = Resource.objects.order_by(order.replace('-', ''))  # Alors on retire le signe négatif
+    else:
+        res = Resource.objects.order_by("-" + order)
+
+    return render(request, 'rsc_mgmt/index.html', {
+        'res': res,
+        'order_type': order,
+        'type': ResourceType.objects.all(),
+        'user': request.session['user']
+    })
 
 
 def detail(request, rid=0):
@@ -34,20 +55,39 @@ def detail(request, rid=0):
 
 # Suppression de ressource
 def delete(request, rid=0):
+    if not is_admin(request.session['user']):
+        return redirect('index')
+
     try:
-        rsc = Resource(pk=rid).delete()
+        Resource(pk=rid).delete()
     except Resource.DoesNotExist:
         err_msg = "La resource n'a pa été trouvée"
 
     return redirect('index')
 
 
-def edit(request, rid=0):
-    return HttpResponse("Edit page, id : %i" % rid)
+def edit(request, rid=None):
+    res = None
+
+    if rid is not None:
+        res = Resource.objects.get(pk=rid)
+
+    form = CreateResourceForm(request.POST or None, instance=res)
+    if request.POST and form.is_valid():
+        form.save()
+
+        return redirect('index')
+
+    return render(request, 'rsc_mgmt/create_resource.html', {
+        'form': form
+    })
 
 
 def create_res(request):
     # Formulaire de création de ressource
+    if not is_admin(request.session['user']):
+        return redirect('index')
+
     if request.method == 'POST':
         form = CreateResourceForm(request.POST)
 
@@ -62,6 +102,9 @@ def create_res(request):
 
 
 def create_res_type(request):
+    if not is_admin(request.session['user']):
+        return redirect('index')
+
     if request.method == 'POST':
         form = CreateResourceTypeForm(request.POST)
 
@@ -79,8 +122,11 @@ def create_res_type(request):
 
 
 def delete_res_type(request, rid=0):
+    if not is_admin(request.session.get('user')):
+        return redirect('index')
+
     try:
-        rsc = ResourceType(pk=rid).delete()
+        ResourceType(pk=rid).delete()
     except ResourceType.DoesNotExist:
         err_msg = "La resource n'a pa été trouvée"
 
@@ -97,7 +143,14 @@ def login(request):
 
             try:
                 user = User.objects.get(username=username, password=mdp)
-                request.session['user'] = user
+                u = {
+                    'username': user.username,
+                    'is_admin': user.is_admin()
+                }
+
+                request.session['user'] = u
+
+                return redirect('index')
             except User.DoesNotExist:
                 return redirect('login')
